added frontend folder

+++ backend/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require("express-async-handler")
 
 const Goal = require("../model/goalsModel")
+const User = require("../model/userModel")
 
 
 // desc: get goals
 // route: GET /api/goals/
 // access: private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -15,13 +16,14 @@ const getGoals = asyncHandler(async (req, res) => {
 // route: POST /api/goals/
 // access: private
 const setGoal = asyncHandler(async (req, res) => {
-    if (!req.body.text){
+    if (!req.body.text) {
         res.status(400)
         throw new Error("Please add a text field")
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -37,6 +39,22 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error("Goal not found")
     }
+
+    // retrieve user from db using user id
+    const user = await User.findById(req.user.id)
+
+    // check for user if user exists
+    if(!user){
+        res.status(401)
+        throw new Error("User does not exist")
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401).json({"goal":goal.user, "user":user.id})
+        throw new Error("user not authorized")
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true
     })
@@ -52,6 +70,22 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error("Goal not found")
+    }
+
+    // retrieve user from db using user id
+    const user = await User.findById(req.user.id)
+
+
+     // check for user if user exists
+    if(!user){
+        res.status(401)
+        throw new Error("User does not exist")
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401).json({"goal":goal.user, "user":user.id})
+        throw new Error("user not authorized")
     }
 
     await Goal.findByIdAndDelete(req.params.id)
